+++ tests/analyses/test_module.py
@@ -5,7 +5,9 @@
 #  SPDX-License-Identifier: LGPL-3.0-or-later
 #
 import ast
+import importlib
 import itertools
+import typing
 from logging import Logger
 from typing import Any, Union, cast
 from unittest.mock import MagicMock
@@ -15,6 +17,7 @@ from ordered_set import OrderedSet
 
 from pynguin.analyses import module
 from pynguin.analyses.module import (
+    MODULE_BLACKLIST,
     ModuleTestCluster,
     TypeInferenceStrategy,
     _ParseResult,
@@ -22,6 +25,7 @@ from pynguin.analyses.module import (
     generate_test_cluster,
     parse_module,
 )
+from pynguin.analyses.types import ClassWrapper
 from pynguin.utils.exceptions import ConstructionFailedException
 from pynguin.utils.generic.genericaccessibleobject import (
     GenericAccessibleObject,
@@ -122,7 +126,7 @@ def test_parse_module_replace_no_annotation_by_any_parameter(
 ):
     function_args = parsed_module_no_any_annotation.syntax_tree.body[1].args
     assert function_args.args[0].annotation.id == "Any"
-    assert function_args.args[1].annotation.id == "Any"
+    assert function_args.args[1].annotation.id == "object"
     assert function_args.args[2].annotation.id == "Any"
 
 
@@ -135,7 +139,7 @@ def test_parse_module_replace_no_annotation_by_any_return(
     faz_func_return = parsed_module_no_any_annotation.syntax_tree.body[4].returns
     assert foo_func_return == "Any"
     assert bar_func_return == "Any"
-    assert baz_func_return == "Any"
+    assert baz_func_return == "object"
     assert isinstance(faz_func_return, ast.Constant)
     assert faz_func_return.value is None
 
@@ -148,7 +152,7 @@ def test_analyse_module(parsed_module_no_dependencies):
 def test_analyse_module_dependencies(parsed_module_complex_dependencies):
     test_cluster = analyse_module(parsed_module_complex_dependencies)
     assert test_cluster.num_accessible_objects_under_test() == 1
-    assert len(test_cluster.generators) == 2
+    assert len(test_cluster.generators) == 3
     assert len(test_cluster.modifiers) == 1
 
 
@@ -253,9 +257,10 @@ def test_select_concrete_type_any(module_test_cluster):
     generator = MagicMock(GenericMethod)
     generator.generated_type.return_value = MagicMock
     module_test_cluster.add_generator(generator)
-    assert module_test_cluster.select_concrete_type(Any) in list(PRIMITIVES) + list(
-        COLLECTIONS
-    ) + [MagicMock]
+    assert (
+        module_test_cluster.select_concrete_type(Any)
+        in list(PRIMITIVES) + list(COLLECTIONS) + MagicMock.mro()
+    )
 
 
 def test_get_all_generatable_types_only_primitive(module_test_cluster):
@@ -268,9 +273,9 @@ def test_get_all_generatable_types(module_test_cluster):
     generator = MagicMock(GenericMethod)
     generator.generated_type.return_value = MagicMock
     module_test_cluster.add_generator(generator)
-    assert module_test_cluster.get_all_generatable_types() == [MagicMock] + list(
-        PRIMITIVES
-    ) + list(COLLECTIONS)
+    assert module_test_cluster.get_all_generatable_types() == OrderedSet(
+        [MagicMock] + list(PRIMITIVES) + list(COLLECTIONS)
+    )
 
 
 def __convert_to_str_count_dict(dic: dict[type, OrderedSet]) -> dict[str, int]:
@@ -302,9 +307,15 @@ def test_nothing_from_blacklist():
     assert cluster.num_accessible_objects_under_test() == 1
 
 
+def test_blacklist_is_valid():
+    # Naive test without assert, checks if the module names are valid.
+    for item in MODULE_BLACKLIST:
+        importlib.import_module(item)
+
+
 def test_nothing_included_multiple_times():
     cluster = generate_test_cluster("tests.fixtures.cluster.diamond_top")
-    assert sum(len(cl) for cl in cluster.generators.values()) == 5
+    assert sum(len(cl) for cl in cluster.generators.values()) == 6
     assert cluster.num_accessible_objects_under_test() == 1
 
 
@@ -312,7 +323,7 @@ def test_generators():
     cluster = generate_test_cluster("tests.fixtures.cluster.no_dependencies")
     assert len(cluster.get_generators_for(int)) == 0
     assert len(cluster.get_generators_for(float)) == 0
-    assert __convert_to_str_count_dict(cluster.generators) == {"Test": 1}
+    assert __convert_to_str_count_dict(cluster.generators) == {"Test": 1, "object": 1}
     assert cluster.num_accessible_objects_under_test() == 4
 
 
@@ -321,6 +332,7 @@ def test_simple_dependencies():
     assert __convert_to_str_count_dict(cluster.generators) == {
         "SomeArgumentType": 1,
         "ConstructMeWithDependency": 1,
+        "object": 1,
     }
     assert cluster.num_accessible_objects_under_test() == 1
 
@@ -328,6 +340,22 @@ def test_simple_dependencies():
 def test_complex_dependencies():
     cluster = generate_test_cluster("tests.fixtures.cluster.complex_dependencies")
     assert cluster.num_accessible_objects_under_test() == 1
+
+
+def test_inheritance_generator():
+    cluster = generate_test_cluster("tests.fixtures.cluster.inheritance")
+    from tests.fixtures.cluster.inheritance import Bar, Foo
+
+    assert len(cluster.get_generators_for(Foo)) == 2
+    assert len(cluster.get_generators_for(Bar)) == 1
+
+
+def test_inheritance_modifier():
+    cluster = generate_test_cluster("tests.fixtures.cluster.inheritance")
+    from tests.fixtures.cluster.inheritance import Bar, Foo
+
+    assert len(cluster.get_modifiers_for(Bar)) == 2
+    assert len(cluster.get_modifiers_for(Foo)) == 1
 
 
 def test_modifier():
@@ -343,7 +371,7 @@ def test_simple_dependencies_only_own_classes():
 def test_resolve_dependencies():
     cluster = generate_test_cluster("tests.fixtures.cluster.typing_parameters")
     assert len(cluster.accessible_objects_under_test) == 3
-    assert len(cluster.generators) == 3
+    assert len(cluster.generators) == 4
 
 
 def test_resolve_optional():
@@ -373,10 +401,7 @@ def test_conditional_import_forward_ref():
     cluster = generate_test_cluster("tests.fixtures.cluster.conditional_import")
     accessible_objects = list(cluster.accessible_objects_under_test)
     constructor = cast(GenericConstructor, accessible_objects[0])
-    assert (
-        str(constructor.inferred_signature.parameters["arg0"])
-        == "<class 'tests.fixtures.cluster.complex_dependency.SomeOtherType'>"
-    )
+    assert constructor.inferred_signature.parameters["arg0"] == typing.Any
 
 
 def test_enums():
@@ -393,11 +418,18 @@ def test_enums():
 
 @pytest.mark.parametrize(
     "module_name",
-    [pytest.param("cluster.comments"), pytest.param("instrumentation.mixed")],
+    ["async_func", "async_gen", "async_class_gen", "async_class_method"],
 )
 def test_analyse_async_function_or_method(module_name):
     with pytest.raises(ValueError):
-        generate_test_cluster(f"tests.fixtures.{module_name}")
+        generate_test_cluster(f"tests.fixtures.cluster.{module_name}")
+
+
+def test_analyse_async_as_dependency():
+    cluster = generate_test_cluster("tests.fixtures.cluster.uses_async_dependency")
+    assert len(cluster.generators) == 4
+    assert len(cluster.modifiers) == 0
+    assert len(cluster.accessible_objects_under_test) == 1
 
 
 def test_import_dependency():
@@ -431,3 +463,18 @@ def test_analyse_empty_enum_module():
         )
     )
     assert len(enums_without_fields) == 0
+
+
+def test_no_abstract_class():
+    cluster = generate_test_cluster("tests.fixtures.cluster.abstract")
+    assert len(cluster.accessible_objects_under_test) == 1
+    assert len(cluster.generators) == 3
+    assert len(cluster.modifiers) == 1
+
+
+def test_inheritance_graph():
+    cluster = generate_test_cluster("tests.fixtures.cluster.inheritance")
+    assert (
+        len(cluster.inheritance_graph.get_subclasses(ClassWrapper.from_type(object)))
+        == 3
+    )