+++ tests/fixtures/cluster/async_class_method.py
@@ -1,0 +1,9 @@
+#  This file is part of Pynguin.
+#
+#  SPDX-FileCopyrightText: 2019–2022 Pynguin Contributors
+#
+#  SPDX-License-Identifier: LGPL-3.0-or-later
+#
+class Foo:
+    async def foo(self):
+        return 45